2006-11-14 [colin]	2.6.0cvs44
	* src/compose.c
	* src/inc.c
	* src/messageview.c
	* src/news.c
	* src/pop.c
	* src/procmsg.c
	* src/gtk/gtkaspell.c
		Fix 64-bit-related warnings
	* src/ssl_manager.c
		Fix possible infinite loop
	* src/plugins/pgpcore/sgpgme.c
		Remove useless check

+++ src/compose.c
@@ -2224,7 +2224,7 @@ static void compose_entries_set(Compose *compose, const gchar *mailto)
 	gchar *subject = NULL;
 	gchar *body = NULL;
 	gchar *temp = NULL;
-	guint  len = 0;
+	gsize  len = 0;
 	gchar *attach = NULL;
 	
 	scan_mailto_url(mailto, &to, &cc, NULL, &subject, &body, &attach);
